requirements.txt
asgiref==3.11.0
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.3
django-cloudinary-storage==0.3.0
django-filter==25.2
djangorestframework==3.17.1
gunicorn==25.1.0
h11==0.16.0
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2==2.9.11
python-dotenv==1.2.1
requests==2.32.5
resend==2.25.0
six==1.17.0
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
whitenoise==6.12.0
